Agregado eliminar registros de usuario_obra_social al eliminar Profesional

+++ app/src/app/components/profesional/list-profesional/list-profesional.component.ts
@@ -166,18 +166,22 @@ export class ListProfesionalComponent  implements OnInit{
      
 
                deleteProfesional(id : number){
-                   this.loading = true;
-                   this._usuarioService.remove(id).subscribe(()=>{
-                        this.loading = false;
-                        this.toastr.success('Usuario eliminado exitosamente', 'Usuario');                        
- 
-                   }, (error)=>{
+                if(confirm('Desea eliminar este registro?')){
+                  this.loading = true;
+                  this._usuarioService.remove(id).subscribe(()=>{
                        this.loading = false;
-                       this.errorServer = error.error?.error || 'Error al eliminar usuario';
-                       console.error(this.errorServer);
-                       this.toastr.error(this.errorServer!, 'Error'); 
-                      
-                   })
+                       this.getListProfesionales();
+                       this.toastr.success('Usuario eliminado exitosamente', 'Usuario');                        
+
+                  }, (error)=>{
+                      this.loading = false;
+                      this.errorServer = error.error?.error || 'Error al eliminar usuario';
+                      console.error(this.errorServer);
+                      this.toastr.error(this.errorServer!, 'Error'); 
+                     
+                  })
+                }
+         
 
                }
